Add outside forbidden zone
