+++ ai/STA/Strategy/penalty_offense.py
@@ -7,7 +7,8 @@ from ai.STA.Strategy.team_go_to_position import TeamGoToPosition
 from ai.STA.Tactic.go_kick import GoKick
 from ai.STA.Tactic.goalkeeper import GoalKeeper
 from ai.states.game_state import GameState
-from Util.geometry import Area
+from Util.area import Area
+
 
 class PenaltyOffense(TeamGoToPosition):
     def __init__(self, game_state):
